Update wheels' brakeInput only when necessary.
If the brake action group is active, then no updating happens as the group
overrides the linear input (to allow for parking brakes), though this was
the previous behavior anyway. However, when it is not active, now the brake
inputs are updated only when there has been a change in the brake values
(ie, when the Update*Braking functions were called) or when the brake group
is deactivated.