[InputLib] Use a device id to find associated C# device
So far, this is just for removing the device, but it means the axis and
button callbacks will be usable as their data parameter will be the
device id.

The device id is just the index in the low-level device list, but slots
are reused rather than holes filled in when a device is removed.